pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<3.11"]
build-backend = "flit_core.buildapi"

[project]
name = "azuremlconstructor"
authors = [{name = "Aziz Nadirov", email = "aziznadirov@yahoo.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License",
                "Development Status :: 4 - Beta",
                "Intended Audience :: Developers",
                "Operating System :: OS Independent"]

dynamic = ["version", "description"]

requires-python = ">=3.9,<3.11"
dependencies = [
    "azureml-core >=1.49.0",
    "azureml-pipeline >=1.49.0",
    "python-dotenv >=1.0.0",
    "pandas >=1.5.3",
    "pyarrow >=9.0.0",
    "pydantic >=1.10.5",
    "Jinja2 >=3.1.2",
    "azure-storage-blob",
]

keywords = ["azure", "machine leaning", "aml", "pipeline"]

[project.urls]
Source = "https://github.com/AzizNadirov/azuremlconstructor"


